fix: [Progress Widgets] Update QTimer and QPropertyAnimation initialization
Updated the initialization of QTimer in the TTKDonutWaitProgressWidget
and TTKZoomWaitProgressWidget classes to ensure proper parent-child
relationship. This change helps in managing the memory of the timer
objects more effectively.

Additionally, modified the QPropertyAnimation initialization in the
TTKRoundProgressWidget to include the parent parameter, improving
resource management.

- Ensures proper memory management for QTimer
- Enhances resource handling for QPropertyAnimation

Log: Update QTimer and QPropertyAnimation initialization